feat(web): add next-auth
