+++ apps/web/next-auth.d.ts
@@ -1,0 +1,1 @@
+import '@types/next-auth';
