+++ apps/web/pages/products.tsx
@@ -2,4 +2,8 @@ import dynamic from 'next/dynamic';
 
 const Products = dynamic(() => import('@shared').then((mod) => mod.Products));
 
+Products.defaultProps = {
+    auth: true,
+};
+
 export default Products;
