+++ apps/web/tsconfig.json
@@ -18,6 +18,6 @@
             "module": "CommonJS"
         }
     },
-    "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx", "next-env.d.ts", "next.config.js"],
+    "include": ["**/*.ts", "**/*.tsx", "**/*.js", "**/*.jsx", "next-env.d.ts", "next-auth.d.ts", "next.config.js"],
     "exclude": ["node_modules", "jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
 }
