+++ libs/shared/setupTests.ts
@@ -1,3 +1,4 @@
 import '@test/lib/mockDynamic';
+import '@test/lib/mockNextAuth';
 import '@testing-library/jest-dom/extend-expect';
 import '@inrupt/jest-jsdom-polyfills';
