+++ libs/shared/src/components/auth/auth.stories.tsx
