+++ libs/shared/src/index.ts
@@ -1,10 +1,10 @@
+export * from './components/auth/auth';
 export * from './components/toaster/toaster';
 export * from './components/toast/toast';
 export * from './pages/products/products';
 export * from './components/product-list/product-list';
 export * from './components/not-signed-in/not-signed-in';
 export * from './components/signed-in/signed-in';
-export * from './pages/auth/auth';
 export * from './components/dropdown/dropdown-menu-radio-item/dropdown-menu-radio-item';
 export * from './components/dropdown/dropdown-menu-content/dropdown-menu-content';
 export * from './components/dropdown/dropdown-menu-sub-trigger/dropdown-menu-sub-trigger';
