+++ libs/shared/src/pages/login/login.spec.tsx
@@ -1,9 +1,28 @@
-import Login from './login';
 import { render } from '@testing-library/react';
+import { LoginProps } from '@types';
+import { getApp } from 'firebase/app';
+import { createUserWithEmailAndPassword, getAuth } from 'firebase/auth';
+import { ComponentType } from 'react';
 
-describe('Login', () => {
+jest.mock('firebase/app');
+jest.mock('firebase/auth');
+
+describe('Auth', () => {
+    let Login: ComponentType<LoginProps>;
+
+    beforeEach(() => {
+        getAuth as jest.Mock;
+        getApp as jest.Mock;
+        createUserWithEmailAndPassword as jest.Mock;
+        Login = require('./login').default as typeof Login;
+    });
+
+    afterAll(() => {
+        jest.resetAllMocks();
+    });
+
     it('should render successfully', () => {
-        const { baseElement } = render(<Login />);
+        const { baseElement } = render(<Login name={''} email={''} password={''} confirmPassword={''} />);
         expect(baseElement).toBeTruthy();
     });
 });
