+++ libs/test/src/lib/mockNextAuth.ts
@@ -1,0 +1,2 @@
+jest.mock('next-auth');
+jest.mock('next-auth/react');
