+++ libs/types/src/lib/auth-props.ts
@@ -1,4 +1,4 @@
-export interface AuthProps {
+export interface LoginProps {
     name: string;
     email: string;
     password: string;
