+++ libs/types/src/lib/next-auth.d.ts
@@ -1,0 +1,21 @@
+import { type DefaultSession } from 'next-auth';
+
+declare module 'next-auth' {
+    /**
+     * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
+     */
+    interface Session {
+        user: {
+            /** The user's postal address. */
+            address: string;
+        } & DefaultSession['user'];
+    }
+}
+
+declare module 'next-auth/jwt' {
+    /** Returned by the `jwt` callback and `getToken`, when using JWT sessions */
+    interface JWT {
+        /** OpenID ID Token */
+        idToken?: string;
+    }
+}
