+++ package.json
@@ -54,6 +54,7 @@
         "lucide-react": "^0.115.0",
         "next": "13.1.1",
         "next-absolute-url": "^1.2.2",
+        "next-auth": "^4.20.1",
         "next-compose-plugins": "^2.2.1",
         "next-firebase-auth": "canary",
         "next-fonts": "^1.5.1",
